Environments and Value Iteration Algorithms
Provides the initial grid world environment and an example of value iteration in both deterministic and stochastic environments.
Also provides an updated version of the environment and a slightly advanced state representation for when there is more than one gold (and therefore varying terminal states). Some of the code in Environment3 is still hard coded to fixed gold positions and only two gold - pending final update.